move types for tracking to top level types
