+++ packages/search/src/search-list/search-list.tsx
@@ -6,7 +6,7 @@ import ArticleListItemSeparator from "@times-components-native/article-list/src/
 import SearchListLoader from "@times-components-native/search/src/search-list/search-list-loader";
 import { styles } from "./styles/search-list-styles";
 import SearchListEmptyState from "./search-list-empty-state";
-import { TTrackingData } from "../types";
+import { TTrackingData } from "@times-components-native/types";
 
 export interface SearchListProps {
   hits: Hit[];
