+++ packages/search/src/types.ts
@@ -86,14 +86,3 @@ export interface Hit {
     };
   };
 }
-
-type TTrackingAtrributes = {
-  eventTime: Date;
-};
-
-export type TTrackingData = {
-  object: string;
-  action: string;
-  component: string;
-  attrs?: TTrackingAtrributes;
-};
