+++ packages/types/tracking.ts
@@ -1,0 +1,10 @@
+type TTrackingAtrributes = {
+  eventTime: Date;
+};
+
+export type TTrackingData = {
+  object: string;
+  action: string;
+  component: string;
+  attrs?: TTrackingAtrributes;
+};
